[pipaudit] Add option to ignore vulnerabilities (#8)
* [pipaudit] Add an option to ignore vulnerabilities

Extend the pipaudit rule, adding an option to ignore vulnerabilities.

Sometimes a vulnerability does not have a fix yet. In such a case, the
user must be able to ignore those as there is no other option for
third-party libraries.

* [security] Ignore `GHSA-4xh5-x5gv-qwph` from `pip`

The vulnerability `GHSA-4xh5-x5gv-qwph` is now affecting pip 25.2 as well but
there is still no fix for it. Hence, it needs to be tolerated for now.

* wip - address review findings