[shell] fix shebang
Make the shebang the first line of run_all_tests, which actually makes
it to have an effect.

Before the shebang was moved to after the copyright header by the S-Core
copyright header check. Since that check has been fixed to accept
something else than the copyright header at the top of the file, the
shebang can be moved to the correct place.