[tests] add workspace modes
Extend run_all_tests to support an optional argument that allows to
select which tests to run, either all, only tests in the main workspace
or only tests in the test workspace.

This shall help to reduce the CI resource usage. In particular the
amount of disk storage needed, if only either of the test workspaces is
executed on a separate CI node.