GitHub action run all tests (#11)
* [shell] fix shebang

Make the shebang the first line of run_all_tests, which actually makes
it to have an effect.

Before the shebang was moved to after the copyright header by the S-Core
copyright header check. Since that check has been fixed to accept
something else than the copyright header at the top of the file, the
shebang can be moved to the correct place.

* [tests] add workspace modes

Extend run_all_tests to support an optional argument that allows to
select which tests to run, either all, only tests in the main workspace
or only tests in the test workspace.

This shall help to reduce the CI resource usage. In particular the
amount of disk storage needed, if only either of the test workspaces is
executed on a separate CI node.

* [ci] Add github test actions

Add github actions to run tests in the main and test workspace.